requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
django-bootstrap-v5==1.0.11
django-bootstrap5==21.3
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-datetime-widget2==0.9.5
django-debug-toolbar==3.5.0
django-dotenv==1.4.2
django-money==3.0.0
django-storages==1.13.1
django-simple-history==3.3.0
pipenv==2022.6.7
Pillow==9.4.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
whitenoise==6.2.0